mcp/requirements.txt
# MCP Server Dependencies
mcp>=1.0.0
neo4j>=5.0.0
pydantic>=2.0.0
pyyaml>=6.0
aiohttp>=3.8.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0 